backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
langchain==0.1.0
langchain-openai==0.0.2
faiss-cpu==1.7.4
numpy==1.24.3
pandas==2.0.3
requests==2.31.0
pydantic==2.4.2
python-dotenv==1.0.0
beautifulsoup4==4.12.2
lxml==4.9.3
chromadb==0.4.18
sentence-transformers==2.2.2
httpx==0.25.2
python-multipart==0.0.6
cors==1.0.1